% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptOptions.R
\name{conceptOptions}
\alias{conceptOptions}
\title{InvestCEE
This function provides access to a list of available concepts stored in a central repository}
\usage{
conceptOptions()
}
\description{
InvestCEE
This function provides access to a list of available concepts stored in a central repository
}
