% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryOptions.R
\name{countryOptions}
\alias{countryOptions}
\title{InvestCEE
This function provides access to the list of available countries stored in a central repository}
\usage{
countryOptions(concept = NULL)
}
\description{
InvestCEE
This function provides access to the list of available countries stored in a central repository
}
\examples{
countryOptions("unemployment") or countryOptions()
}
