% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCode.R
\name{getCode}
\alias{getCode}
\title{InvestCEE
This function provides access to the a specific data code stored in a central repository}
\usage{
getCode(conceptCode, countryCode)
}
\description{
InvestCEE
This function provides access to the a specific data code stored in a central repository
}
