% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescription.R
\name{getDescription}
\alias{getDescription}
\title{InvestCEE
This function provides access to the a specific data description stored in a central repository}
\usage{
getDescription(conceptCode, countryCode)
}
\description{
InvestCEE
This function provides access to the a specific data description stored in a central repository
}
