% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFrequency.R
\name{getFrequency}
\alias{getFrequency}
\title{InvestCEE
This function provides access to the a specific data frequency stored in a central repository}
\usage{
getFrequency(conceptCode, countryCode, frequency)
}
\description{
InvestCEE
This function provides access to the a specific data frequency stored in a central repository
}
