% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProvider.R
\name{getProvider}
\alias{getProvider}
\title{InvestCEE
This function provides access to the a specific data provider stored in a central repository}
\usage{
getProvider(conceptCode, countryCode, provider)
}
\description{
InvestCEE
This function provides access to the a specific data provider stored in a central repository
}
