% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_invcee.R
\name{theme_invcee}
\alias{theme_invcee}
\title{Theme InvestCEE
This function provides the look of ggplots for InvestCEE.com
Based on theme_grey, it sets font, background, legend position,
axis, gridlines, with blue background}
\usage{
theme_invcee()
}
\description{
Theme InvestCEE
This function provides the look of ggplots for InvestCEE.com
Based on theme_grey, it sets font, background, legend position,
axis, gridlines, with blue background
}
