% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_invcee_gray.R
\name{theme_invcee_gray}
\alias{theme_invcee_gray}
\title{Theme InvestCEE
This function provides the look of ggplots for InvestCEE.com
Based on theme_minimal, it sets font, background, legend position,
axis, gridlines, with gray background}
\usage{
theme_invcee_gray()
}
\description{
Theme InvestCEE
This function provides the look of ggplots for InvestCEE.com
Based on theme_minimal, it sets font, background, legend position,
axis, gridlines, with gray background
}
