% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_invcee_plain.R
\name{theme_invcee_plain}
\alias{theme_invcee_plain}
\title{Theme InvestCEE. Special adjustments for plain look
This function provides the look of ggplots for InvestCEE.com
Based on theme_grey, it sets font, background, legend position,
axis, gridlines, with white background}
\usage{
theme_invcee_plain()
}
\description{
Theme InvestCEE. Special adjustments for plain look
This function provides the look of ggplots for InvestCEE.com
Based on theme_grey, it sets font, background, legend position,
axis, gridlines, with white background
}
