% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_invcee.R
\name{theme_jack}
\alias{theme_jack}
\title{Theme InvestCEE
This function provides the look of ggplots for InvestCEE.com
Based on theme_minimal, it sets font, background, legend position,
axis, gridlines}
\usage{
theme_jack()
}
\description{
Theme InvestCEE
This function provides the look of ggplots for InvestCEE.com
Based on theme_minimal, it sets font, background, legend position,
axis, gridlines
}
